+++ search.go
@@ -81,14 +81,6 @@ func newSearchFilterFromQuery(query url.Values) (*packages.Filter, error) {
 		}
 	}
 
-	if v := query.Get("internal"); v != "" {
-		// In case of error, keep it false
-		filter.Internal, err = strconv.ParseBool(v)
-		if err != nil {
-			return nil, fmt.Errorf("invalid 'internal' query param: '%s'", v)
-		}
-	}
-
 	if v := query.Get("experimental"); v != "" {
 		// In case of error, keep it false
 		filter.Experimental, err = strconv.ParseBool(v)